Modified encryption folder
