+++ src/app/encryption/hash-pass.service.ts
@@ -1,0 +1,13 @@
+import { Injectable } from '@angular/core';
+import { CryptoJS } from 'crypto-js';
+
+@Injectable()
+export class HashPassService {
+
+  constructor() { }
+
+  public hashPassword(password: string) {
+    let hash = CryptoJS.SHA256(password);
+    return '';
+  }
+}
