--- src/app/encryption/password-hash.service.ts
@@ -1,13 +1,0 @@
-import { Injectable } from '@angular/core';
-
-@Injectable()
-export class PasswordHashService {
-
-  constructor() { }
-
-  public hashPassword(password: string): string {
-
-    return '';
-  }
-
-}
